Adding raspeedi data editing
