+++ dashboard/templates/dashboard/nav_csd_atelier.html
@@ -28,7 +28,6 @@
                     <h6 class="collapse-header">Edition :</h6>
                     <a class="collapse-item" href="{% url 'squalaetp:corvet_insert' %}">{% trans "Add Corvet" %}</a>
                     <a class="collapse-item" href="#">{% trans "Add data" %}</a>
-                    <a class="collapse-item" href="#">{% trans "Edit data" %}</a>
                 </div>
             </div>
         </li>
@@ -53,7 +52,6 @@
                     <div class="collapse-divider"></div>
                     <h6 class="collapse-header">Edition :</h6>
                     <a class="collapse-item" href="{% url 'raspeedi:insert' %}">{% trans "Add product" %}</a>
-                    <a class="collapse-item" href="#">{% trans "Edit product" %}</a>
                 </div>
             </div>
         </li>
