+++ raspeedi/locale/fr/LC_MESSAGES/django.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-06-18 12:38+0200\n"
+"POT-Creation-Date: 2019-06-25 10:19+0200\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -66,7 +66,7 @@ msgstr "Référence MM"
 msgid "Submit"
 msgstr "Soumettre"
 
-#: raspeedi/tests/raspeedi_views.py:51
+#: raspeedi/tests/tests_raspeedi_views.py:47
 msgid "This field is required."
 msgstr ""
