+++ raspeedi/templates/raspeedi/forms.html
@@ -1,0 +1,106 @@
+{% load i18n %}
+
+{% csrf_token %}
+
+<div class="form-group row">
+    <div class="col-sm-6 mb-3 mb-sm-0">
+        <label for="{{ form.ref_boitier.id_for_label }}">{% trans "case reference" %} : <span
+                class="text-danger">*</span></label>
+
+        {{ form.ref_boitier }}
+
+    </div>
+    <div class="col-sm-3">
+        <label for="{{ form.produit.id_for_label }}">{% trans "Product" %} : <span
+                class="text-danger">*</span></label>
+
+        {{ form.produit }}
+
+    </div>
+    <div class="col-sm-3">
+        <label for="{{ form.facade.id_for_label }}">{% trans "Front" %} : <span class="text-danger">*</span></label>
+
+        {{ form.facade }}
+
+    </div>
+</div>
+<div class="form-group row">
+    <div class="col-sm-6 mb-3 mb-sm-0">
+        <label for="{{ form.type.id_for_label }}">Type : <span class="text-danger">*</span></label>
+
+        {{ form.type }}
+
+    </div>
+    <div class="col-sm-3">
+        <label for="{{ form.dab.id_for_label }}">DAB :</label>
+
+        {{ form.dab }}
+
+    </div>
+    <div class="col-sm-3">
+        <label for="{{ form.cam.id_for_label }}">{% trans "Reversing camera" %} :</label>
+
+        {{ form.cam }}
+
+    </div>
+</div>
+<div class="form-group row">
+    <div class="col-sm-6 mb-3 mb-sm-0">
+        <label for="{{ form.dump_peedi.id_for_label }}">{% trans "PEEDI dump" %} :</label>
+
+        {{ form.dump_peedi }}
+    </div>
+    <div class="col-sm-6">
+        <label for="{{ form.dump_renesas.id_for_label }}">{% trans "RENESAS dump" %} :</label>
+
+        {{ form.dump_renesas }}
+    </div>
+</div>
+<div class="form-group row">
+    <div class="col-sm-3 mb-3 mb-sm-0">
+        <label for="{{ form.cd_version.id_for_label }}">{% trans "Version CD" %} :</label>
+
+        {{ form.cd_version }}
+
+    </div>
+    <div class="col-sm-3">
+        <label for="{{ form.media.id_for_label }}">{% trans "Media type" %} : <span
+                class="text-danger">*</span></label>
+
+        {{ form.media }}
+
+    </div>
+    <div class="col-sm-3">
+        <label for="{{ form.carto.id_for_label }}">{% trans "Carto version" %} :</label>
+
+        {{ form.carto }}
+
+    </div>
+    <div class="col-sm-3">
+        <label for="{{ form.connecteur_ecran.id_for_label }}">{% trans "Screens number" %} : <span
+                class="text-danger">*</span></label>
+
+        {{ form.connecteur_ecran }}
+
+    </div>
+</div>
+<div class="form-group row">
+    <div class="col-sm-6 mb-3 mb-sm-0">
+        <label for="{{ form.ref_mm.id_for_label }}">{% trans "MM reference" %} :</label>
+
+        {{ form.ref_mm }}
+    </div>
+</div>
+
+{% if errors %}
+    {% for key, error in errors %}
+        {% autoescape off %}
+            {{ error }}
+        {% endautoescape %}
+    {% endfor %}
+{% endif %}
+
+<button type="submit" name="btn_raspeedi_insert"
+        class="col-sm-4 offset-sm-4 align-self-center btn btn-primary btn-block">
+    {% trans "Submit" %}
+</button>
