+++ raspeedi/templates/raspeedi/table.html
@@ -53,7 +53,8 @@
                 <td>{{ prod.media }}</td>
                 <td>{{ prod.carto }}</td>
                 <td>{{ prod.dump_renesas }}</td>
-                <td><a href="#" class="btn btn-success btn-circle btn-sm"><i class="fas fa-edit"></i></a></td>
+                <td><a href="{% url 'raspeedi:edit' ref_case=prod.ref_boitier %}"
+                       class="btn btn-success btn-circle btn-sm"><i class="fas fa-edit"></i></a></td>
                 <td><a href="#" class="btn btn-info btn-circle btn-sm"><i class="fas fa-info-circle"></i></a></td>
             </tr>
