+++ raspeedi/urls.py
@@ -7,4 +7,5 @@ app_name = 'raspeedi'
 urlpatterns = [
     path('table/', views.table, name='table'),
     path('insert/', views.insert, name='insert'),
+    path('<int:ref_case>/edit/', views.edit, name='edit'),
 ]
