+++ sbadmin/locale/fr/LC_MESSAGES/django.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-06-18 12:38+0200\n"
+"POT-Creation-Date: 2019-06-25 10:19+0200\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -18,10 +18,10 @@ msgstr ""
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n > 1);\n"
 
-#: sbadmin/settings/__init__.py:125
+#: sbadmin/settings/__init__.py:140
 msgid "French"
 msgstr "Français"
 
-#: sbadmin/settings/__init__.py:126
+#: sbadmin/settings/__init__.py:141
 msgid "English"
 msgstr "Anglais"
